+++ fj-doc-playground-quarkus/src/main/java/org/fugerit/java/doc/playground/val/ValRest.java
@@ -9,7 +9,7 @@ import org.fugerit.java.doc.val.core.DocValidatorFacade;
 import org.fugerit.java.doc.val.core.basic.ImageValidator;
 import org.fugerit.java.doc.val.core.basic.XmlValidator;
 import org.fugerit.java.doc.val.p7m.P7MValidator;
-import org.fugerit.java.doc.val.pdf.box.PdfboxValidator;
+import org.fugerit.java.doc.val.pdf.box3.Pdfbox3Validator;
 import org.fugerit.java.doc.val.poi.DocValidator;
 import org.fugerit.java.doc.val.poi.DocxValidator;
 import org.fugerit.java.doc.val.poi.XlsValidator;
@@ -32,7 +32,7 @@ import lombok.extern.slf4j.Slf4j;
 public class ValRest {
 
     private static final DocValidatorFacade facade = DocValidatorFacade.newFacadeStrict(ImageValidator.JPG_VALIDATOR,
-            ImageValidator.PNG_VALIDATOR, ImageValidator.TIFF_VALIDATOR, PdfboxValidator.DEFAULT, DocxValidator.DEFAULT,
+            ImageValidator.PNG_VALIDATOR, ImageValidator.TIFF_VALIDATOR, Pdfbox3Validator.DEFAULT, DocxValidator.DEFAULT,
             DocValidator.DEFAULT, XlsxValidator.DEFAULT, XlsValidator.DEFAULT, P7MValidator.DEFAULT, XmlValidator.DEFAULT);
 
     @POST