+++ fj-doc-val-pdfbox3/src/test/java/test/org/fugerit/java/doc/pdfbox3/val/TestPdfbox3Validator.java
@@ -1,0 +1,32 @@
+package test.org.fugerit.java.doc.pdfbox3.val;
+
+import org.fugerit.java.doc.val.core.DocValidatorFacade;
+import org.fugerit.java.doc.val.pdf.box3.Pdfbox3Validator;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+class TestPdfbox3Validator extends TestDocValidatorFacade {
+
+	private static final DocValidatorFacade FACADE = DocValidatorFacade.newFacadeStrict( 
+			Pdfbox3Validator.DEFAULT
+	);
+	
+	@Test
+	void testPdfAsPdf() {
+		boolean ok = this.worker(FACADE, "pdf_as_pdf.pdf", true );
+		Assertions.assertTrue( ok );
+	}
+	
+	@Test
+	void testPngAsPdf() {
+		boolean ok = this.worker(FACADE, "png_as_pdf.pdf", false );
+		Assertions.assertTrue( ok );
+	}
+	
+	@Test
+	void testPdfAsPdfP7M() {
+		boolean ok = this.worker(FACADE, "pdf_as_pdf_p7m.pdf", true );
+		Assertions.assertTrue( ok );
+	}
+	
+}